Move sketch.js into components folder
 and do some cleanup
